+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## v1.1.2
+- Changed formatting time from minutes to seconds
+- Show only time when show percentage is disabled and show time is enabled
+- Made sure progress does not go over 100%
+- Added option to log the changes to hover text
+- Changed settings: Renamed "Show Color Percentage" to "Show Percentage Color". Renamed "Amount of Decimal Places" to "Show Percentage Decimal Places"
+- Attempt to read deprecated settings and remove them (NOT WORKING)
 ## v1.1.0
 - Created new project from bepinex5plugin template
 - Added option to show time remaining